% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_treatment_effects.R
\name{print.individual_treatment_effects}
\alias{print.individual_treatment_effects}
\title{Printing individualized treatment effects}
\usage{
\method{print}{individual_treatment_effects}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{a fitted object from either \code{\link[personalized]{treat.effects}} or \code{\link[personalized]{treatment.effects}}}

\item{digits}{minimal number of significant digits to print.}

\item{...}{further arguments passed to or from \code{\link[base]{print.default}}.}
}
\description{
Prints results for estimated subgroup treatment effects
}
