% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_subgroup.R
\name{print.subgroup_fitted}
\alias{print.subgroup_fitted}
\alias{print.subgroup_validated}
\title{Printing results for fitted subgroup identification models}
\usage{
\method{print}{subgroup_fitted}(x, digits = max(getOption("digits") - 3, 3),
  ...)

\method{print}{subgroup_validated}(x, digits = max(getOption("digits") - 3,
  3), ...)
}
\arguments{
\item{x}{a fitted object from either \code{fit.subgroup} or \code{validate.subgroup}}

\item{digits}{minimal number of significant digits to print.}

\item{...}{further arguments passed to or from \code{\link[base]{print.default}}.}
}
\description{
Prints results for estimated subgroup treatment effects
}
\seealso{
\code{\link[personalized]{validate.subgroup}} for function which creates validation results
and \code{\link[personalized]{fit.subgroup}} for function which fits subgroup identification models.
}
