% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_subgroups.R
\name{summarize.subgroups}
\alias{summarize.subgroups}
\title{Summarizing covariates within estimated subgroups}
\usage{
summarize.subgroups(object)
}
\arguments{
\item{object}{a fitted object from \code{fit.subgroup()}}
}
\description{
Summarizes covariate values within the estimated subgroups
}
\seealso{
\code{\link[personalized]{fit.subgroup}} for function which fits subgroup identification models.
}
